+++ js/downloadbutton/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/downloadbutton
 
+## 0.4.15
+
+### Dependency updates
+
+- @gradio/utils@0.11.0
+- @gradio/client@2.0.1
+- @gradio/button@0.6.1
+
 ## 0.4.14
 
 ### Dependency updates