+++ js/state/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gradio/state
 
+## 0.2.1
+
+### Dependency updates
+
+- @gradio/utils@0.11.0
+
 ## 0.2.0
 
 ### Dependency updates